Prebuild list error state
