+++ components/dashboard/src/prebuilds/list/PrebuildListErrorState.tsx
@@ -8,14 +8,15 @@ import { FC } from "react";
 import { Heading2, Subheading } from "@podkit/typography/Headings";
 import { cn } from "@podkit/lib/cn";
 
-export const PrebuildListEmptyState: FC = () => {
+type Props = {
+    error: unknown;
+};
+export const PrebuildListErrorState: FC<Props> = ({ error }: Props) => {
     return (
-        <div className={cn("w-full flex justify-center mt-2 rounded-xl bg-gray-100 dark:bg-gray-800 px-4 py-20")}>
+        <div className={cn("w-full flex justify-center mt-2 rounded-xl bg-pk-surface-secondary px-4 py-20")}>
             <div className="flex flex-col justify-center items-center text-center space-y-4">
-                <Heading2>No prebuilds yet</Heading2>
-                <Subheading className="max-w-md">
-                    Go on, create Configurations and turn them on, I'll be waiting here...
-                </Subheading>
+                <Heading2>Prebuilds failed to load</Heading2>
+                <Subheading className="max-w-md">Error: {error}.</Subheading>
             </div>
         </div>
     );
